// src/app/api/model/badgeRequest.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BadgeRequest { 
    id?: number;
    nom: string;
    description: string;
    type: BadgeRequest.TypeEnum;
    icone: string;
}
export namespace BadgeRequest {
    export const TypeEnum = {
        Or: 'OR',
        Argent: 'ARGENT',
        Bronze: 'BRONZE',
        Special: 'SPECIAL'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}